Added scrapegraph's RAG node to the CodeGeneratorGraph
ChatGPT was complaining about scraping eviction data in `test-5` because there wasn't any to scrape on the search page - i was hoping it would look at the page and plug some default values. So I've changed it to scraping zillow since it's a lot harder.

There are now default options for LLM, you can control temperature, top_p, max_tokens, and seed. More testing is required with temperature but setting it to 0 is not looking amazing since the consistent results have been mediocre.